投稿機能に関するindex.jsのコードをpost.jsに移動
